fix(exchange): add some nonce checks
